apps/api/requirements.txt
# ══════════════════════════════════════════════════════════════════════════════
# AUJASYA — Python Dependencies (Pinned)
# [FIX-7] PyJWT replaces python-jose (abandoned, CVEs)
# [FIX-17] pydantic-settings is separate from pydantic v2
# ══════════════════════════════════════════════════════════════════════════════

# ── Framework ─────────────────────────────────────────────────────────────────
fastapi>=0.111.0,<0.112.0
uvicorn[standard]>=0.30.0,<0.31.0
gunicorn>=22.0.0,<23.0.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.30,<2.1.0
asyncpg>=0.29.0,<0.30.0
psycopg2-binary>=2.9.9,<2.10.0
alembic>=1.13.0,<1.14.0

# ── Validation ────────────────────────────────────────────────────────────────
pydantic>=2.7.0,<2.8.0
pydantic-settings>=2.2.0,<3.0.0

# ── Cache / Sessions ─────────────────────────────────────────────────────────
redis[hiredis]>=5.0.0,<6.0.0

# ── Background Tasks ─────────────────────────────────────────────────────────
celery>=5.4.0,<5.5.0

# ── Authentication [FIX-7] ────────────────────────────────────────────────────
PyJWT>=2.8.0,<3.0.0
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt>=4.1.0,<5.0.0

# ── Encryption ────────────────────────────────────────────────────────────────
cryptography>=42.0.0,<43.0.0

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx>=0.27.0,<0.28.0

# ── Firebase ──────────────────────────────────────────────────────────────────
firebase-admin>=6.5.0,<7.0.0

# ── Monitoring ────────────────────────────────────────────────────────────────
sentry-sdk[fastapi]>=2.0.0,<3.0.0
structlog>=24.1.0,<25.0.0

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0,<2.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.23.0,<0.24.0
pytest-cov>=5.0.0,<6.0.0

# ── Linting & Type Checking ──────────────────────────────────────────────────
ruff>=0.4.0,<0.5.0
mypy>=1.10.0,<2.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-multipart>=0.0.9,<0.1.0
